+++ src/language/codeQuality.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode';
 import { BATCH_MODE, PSL_MODE, TRIG_MODE } from '../extension';
-import * as parser from '../parser/parser';
+import * as parser from 'psl-parser';
 import { getDiagnostics } from '../pslLint/activate';
 import * as api from '../pslLint/api';
 import { getConfig, removeConfig, setConfig } from '../pslLint/config';
