+++ src/language/pslSignature.ts
@@ -1,7 +1,7 @@
 import * as vscode from 'vscode';
 import { FinderPaths, getFinderPaths } from '../parser/config';
-import * as parser from '../parser/parser';
-import { Position, Token } from '../parser/tokenizer';
+import * as parser from 'psl-parser';
+import { Position, Token } from 'psl-parser';
 import * as utils from '../parser/utilities';
 import * as lang from './lang';
