+++ src/language/pslSuggest.ts
@@ -1,8 +1,8 @@
 import * as path from 'path';
 import * as vscode from 'vscode';
 import { FinderPaths, getFinderPaths } from '../parser/config';
-import * as parser from '../parser/parser';
-import { MemberClass } from '../parser/parser';
+import * as parser from 'psl-parser';
+import { MemberClass } from 'psl-parser';
 import * as utils from '../parser/utilities';
 import * as lang from './lang';
