--- src/parser/index.ts
@@ -1,9 +1,0 @@
-export * from './parser';
-
-export * from './statementParser';
-
-import * as _tokenizer from './tokenizer';
-export { _tokenizer as tokenizer };
-
-import * as _utilities from './utilities';
-export { _utilities as utilities };
