+++ src/pslLint/api.ts
@@ -1,6 +1,6 @@
 import * as path from 'path';
-import { Declaration, Member, Method, Parameter, ParsedDocument, Property } from './../parser/parser';
-import { Position, Range } from './../parser/tokenizer';
+import { Declaration, Member, Method, Parameter, ParsedDocument, Property } from 'psl-parser';
+import { Position, Range } from 'psl-parser';
 
 export enum DiagnosticSeverity {
