+++ src/pslLint/runtime.ts
@@ -1,9 +1,9 @@
-import { Member, MemberClass, Method } from '../parser/parser';
+import { Member, MemberClass, Method } from 'psl-parser';
 import {
 	BinaryOperator, Identifier,
 	StringLiteral, SyntaxKind, Value,
-} from '../parser/statementParser';
-import { Range, Token } from '../parser/tokenizer';
+} from 'psl-parser';
+import { Range, Token } from 'psl-parser';
 import { getCommentsOnLine } from '../parser/utilities';
 import { Diagnostic, DiagnosticRelatedInformation, DiagnosticSeverity, MethodRule } from './api';
